Implemented the cache for episode metadata, although it's a little bit messy and should be cleaned up a bit.
I've discovered a nasty side effect of my metadata discovery schemes.  One Punch Man is not found in the TVDB (they call it One-Punch Man, and that's apparently really far from One Punch Man), so I search the movie databases for it, and IMDB returns a hit.  So it's now a movie and has an IMDB ID.  That's enough to look for it in the episode_meta table even though we don't have a TVDB for it.

I need to think of solutions, but my natural inclination is add an IMDB search in case tvdb fails to avoid such problems.  I can confirm its media type by examining the 'Type' key, and if it's "series", then I can pass the IMDB ID to the GetSeriesByRemoteID API (http://thetvdb.com/wiki/index.php?title=API:GetSeriesByRemoteID) to get the TVDB ID and retry.  This will prevent any IMDB movie searches from turning up TV show hits.