This commit is a little bit weird, so bear with me.
First, I wrote a class for benchmarking called TimeStamper.  Users can call stamp a bunch of times, and then finish with stamp_and_dump to dump all the timestamps taken with their own messages.  I also added support for an extra time before the timestamper object is initialized (see default.py).

Because I want to use the time stamping for benchmarking, it should be printing things when debugging is not enabled, since I don't want to incur the overhead of the millions of debug statements (TM) in this project.  To overcome that, I added a debug-timestamp setting that can be switched on in the settings, which is useful for turning them on or off, regardless of whether debug logging is turned on.

I then benchmarked my code on the Fire Stick and found that importing everything in controller was adding a lot of extra time to the local menu (controller is always imported by default, which is the only entry point), not to mention blowing up the debug log (looking at you, urlresolver).  One advantage to importing everything at the top of controller was that I could get effectively compilation errors like in compiled languages.

So, I added another debug setting called debug-import that imports everything if enabled.

It's not captured in this commit, but I timed most of the imports to find out that the heavy hitters are nethelpers, loose_metadata, and metadata_finder, and that all of the list types are pretty large too.  The result of this experiment was to move these imports to the necessary functions or the class's initialization function so that these costs are only incurred when necessary.

I really don't like the way it works, and it only minimally impacted perf on a beefy Windows machine, but the perf gains on the Fire Stick for the main menu are undeniable: the time it took to access the main menu dropped from 4 seconds to 0.4 seconds!

Other changes include:
 - Got rid of orphan imports
 - Bug fix in EpisodeList.parse and add_items: the bookmark ID is not present if the user is not logged in.

Next up