Found out how KissAnime handles logged in sessions.  I already knew that they use username+password cookies, but I didn't know the expiration date for them.  It turns out they last for 30 years.  Additionally, the username and password cookie values can be used at other computers.
I could have stuck with the original implementation (storing the username and password in plaintext in the settings), but I feel a little bit more comfortable using KissAnime's security system, which admittedly is barely more secure than plaintext.  At least it's obfuscated!

The new implementation asks the user for the username and password and then immediately gets the cookies and updates the setting, not storing the username and password anywhere.  I strictly depend on the cookies for credentials, in other words.

Because the settings are kinda static, I'm sticking to a single button that contains the username if logged in and nothing if logged out.  I guess it would technically be possible to substitute the username for "Log in" and "Log out", but I like showing the username and a little bit more work would be needed to ensure the right value is displayed at start-up, since settings can be changed outside of the plugin.

Other changes include:
 - Added the hidden parameter to get_user_input
 - Fixed a perf bug in Controller.show_bookmark_list where I instantiated the BookmarkList in the account_operation invocation instead of just passing a reference to it
 - Fixed a metahandlers bug where TMDB would error out when trying to format the metadata if runtime's value was None, since that can't be turned into an int.  This caused problems when trying to iterate through a list of metadata and one of the items was just None.

NOTE: THIS IS THE BASE FOR THE FIRST RELEASE.  NEXT UP IS TESTING AND MORE BENCHMARKING!